Add: `formatRolesForForm` permissions helper
